Add application variants and Mbed OS bare metal tutorial
The example project has been re-written to include variants of an application
to blink an LED. The variants include a simple variant that uses a delay within
a loop and a more complex that creates threads and shows inter-thread
communication.
As the simpler variant is single-threaded, it showcases that some of Mbed OS
APIs can be used both with and without rtos (using Mbed OS bare metal profile).
Instructions are also provided to build the simple blinky application with
Mbed OS bare metal profile.

Runtime statistics has been removed and is only mentioned in the README. The
reader is invited to find out more. Runtime statistics is a topic on its own
and should not be implemented here.

The example project also exposes the user to Mbed OS parameter configuration
option.