Simplify application (#192)
The example project has been re-written to be simplified. It uses a delay within
a loop. 

Runtime statistics has been removed and is only mentioned in the README. The
reader is invited to find out more. Runtime statistics is a topic on its own
and should not be implemented here.